Move structs to own files
