+++ src/commands/publish/upload_form/file.rs
@@ -1,0 +1,9 @@
+#[derive(Debug)]
+pub struct File {
+    pub name: String,
+    pub path: String,
+}
+
+pub trait ToFile {
+    fn to_file(&self) -> File;
+}
