+++ src/commands/publish/upload_form/script.rs
@@ -1,0 +1,15 @@
+use super::file::{File, ToFile};
+
+#[derive(Debug)]
+pub struct Script {
+    pub path: String,
+}
+
+impl ToFile for Script {
+    fn to_file(&self) -> File {
+        File {
+            name: "script".to_string(),
+            path: self.path.clone(),
+        }
+    }
+}
